# Use rtree for ring ring checks in polygon validation