Perf improvements:

```
benchmark                                                 old ns/op     new ns/op     delta
BenchmarkIntersectsLineStringWithLineString/n=10-4        2380          2426          +1.93%
BenchmarkIntersectsLineStringWithLineString/n=100-4       20313         19499         -4.01%
BenchmarkIntersectsLineStringWithLineString/n=1000-4      228623        226739        -0.82%
BenchmarkIntersectsLineStringWithLineString/n=10000-4     2468048       2633405       +6.70%
BenchmarkPolygonSingleRingValidation/n=10-4               1905          1844          -3.20%
BenchmarkPolygonSingleRingValidation/n=100-4              18376         19047         +3.65%
BenchmarkPolygonSingleRingValidation/n=1000-4             226838        233961        +3.14%
BenchmarkPolygonSingleRingValidation/n=10000-4            2613638       2645092       +1.20%
BenchmarkPolygonMultipleRingsValidation/n=4-4             20265         17037         -15.93%
BenchmarkPolygonMultipleRingsValidation/n=36-4            340655        157102        -53.88%
BenchmarkPolygonMultipleRingsValidation/n=400-4           9485533       1963860       -79.30%
BenchmarkPolygonMultipleRingsValidation/n=4096-4          301652252     23110987      -92.34%
BenchmarkMultipolygonValidation/n=1-4                     336           336           +0.00%
BenchmarkMultipolygonValidation/n=4-4                     13552         13451         -0.75%
BenchmarkMultipolygonValidation/n=16-4                    154485        154016        -0.30%
BenchmarkMultipolygonValidation/n=64-4                    1434364       1417629       -1.17%
BenchmarkMultipolygonValidation/n=256-4                   12166675      12075248      -0.75%
BenchmarkMultipolygonValidation/n=1024-4                  102995343     103077746     +0.08%
```